requirements/prod.txt
aiohttp==3.9.5
openai==1.28.0
python-socketio==5.11.2
